+++ core/lib/init.js
